feat : view 컨트롤러와 rest 컨트롤러 분리
